Update reamde with deprecation notice
For some time now we have effectively treated the address deduplicator
as deprecated. It technically works but in practice is hard to use
effectively, and query-time deduplication in the API has gotten much
better and is more flexible.

I'd like to eventually archie this Github repositry, and possibly even
mark the NPM packages for the address deduplicator as deprecated. This
update is at least a start of that.

Connects https://github.com/pelias/pelias/issues/702